# Conflicts:
#	src/apps/weblib/cinterface.cpp
#	src/base/conv/auto_json.h
#	src/chart/generator/marker.cpp
#	src/chart/main/events.h
#	src/chart/rendering/drawinterlacing.cpp
#	src/chart/rendering/drawlabel.cpp
#	src/chart/rendering/drawlabel.h
#	src/chart/ui/chart.cpp
#	src/chart/ui/chart.h
#	test/unit/base/conv/auto_json.cpp

+++ src/apps/weblib/jscriptcanvas.cpp
@@ -128,7 +128,7 @@ void JScriptCanvas::beginDropShadow()
 	::canvas_beginDropShadow();
 }
 
-void JScriptCanvas::setDropShadowBlur(uint64_t radius)
+void JScriptCanvas::setDropShadowBlur(double radius)
 {
 	::canvas_setDropShadowBlur(radius);
 }
