+++ src/apps/weblib/jscriptcanvas.h
@@ -30,7 +30,7 @@ public:
 	void setTextColor(const Gfx::Color &color) override;
 
 	void beginDropShadow() override;
-	void setDropShadowBlur(uint64_t radius) override;
+	void setDropShadowBlur(double radius) override;
 	void setDropShadowColor(const Gfx::Color &color) override;
 	void setDropShadowOffset(const Geom::Point &offset) override;
 	void endDropShadow() override;
